refactored animated view for collapse to it's own component
